docs/configuration: fix/update abstract_{chars_count,delimiter}
The description of abstract_delimiter was obviously wrong, copy-pasted
from abstract_chars_count. Fortunately, even if someone didn't notice
this, setting it to a numeric value erred out, therefore nobody can have
such a working configuration and there's no need to complicate things
trying to accept numbers. The best we can do is document the present
behaviour.

In order to make the interaction of these two settings clear, describe
the full mechanism at once and refer to it from the other place.